Change infection logic
Previously, whether a susceptible particle would get infected or not was calculated with a transmission probability on each interaction with infected particles, but two particles tend to interact many times when close, not just once.
Now, each susceptible particle will get infected or not considering the amount of different unique infected particles they encounter each day, and not considering mutiple interactions with each of those.